backend/app/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
pillow==10.3.0
ultralytics==8.1.119
numpy==1.26.0
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
sqlalchemy==2.0.23
authlib==1.2.1
python-dotenv==1.0.0
